Move pagefault handler into pagetable.c
